requirements.txt
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
opencv-python==4.7.0.72
matplotlib==3.7.1
Pillow==9.5.0
scikit-image==0.20.0
tqdm==4.65.0
open3d==0.18.0
timm==0.9.16
streamlit==1.36.0
plotly==5.24.1
